Disable memory thrink opt when memory.size opcode is found
The memory thrink opt is applied when no memory.grow opcode is found
and __heap_base, __data_end and aux_stack are resolved. But the wasm
app can use memory.size opcode to calculate the load/store offset,
e.g. bytecodes generated by go compiler:

    i32.const 65640
    i32.const 0
    memory.size
    i32.const 16
    i32.shl
    i32.const 65640
    i32.sub
    i32.const 6
    i32.shr_u
    call $memset

If we apply the opt, the num_bytes_per_page is changed and the offset
calculated might be out of linear memory. So we disable the opt if
 memory.size opcode is found.

Change-Id: If67ef64effd3ae77e8671696ed1baf738ff868b6