Fix possible malloc(0) and free(NULL) issues
In some platforms, malloc memory with size 0 returns NULL and causes
runtime run failed. We try to avoid allocating memory when the size
is 0.
Adjust size to 1 if it is 0 in wasm_runtime_malloc, and log warning.
If wasm_runtime_free with NULL, also log warning. If warnings are
found, we should fix the related issues.

Change-Id: Idbbe632ea545110ec6196f8d18a9ee345a518907